Add files via upload
Data files needed to run the script of the main process.
Warning. values in 'extraction checkpoint.json' json file should be initialized to empty string (""), in Python, when a brand new process is executed, i.e. there wasn't any unexpected stop that lead the program to end of execution and values in json file were reported for continuing.